Initial Commit
Wine prediction using linear regression is a statistical modeling technique that aims to estimate the quality of wines based on various input features. In this approach, the relationship between the dependent variable (quality) and independent variables (attributes like acidity, sugar levels, pH, etc.) is modeled using a linear equation. The linear regression model is trained on a dataset of wines with known good_quality .we describe this column from quality , and it learns the coefficients that best fit the data. Once trained, the model can then be used to predict the quality of new, unseen wines based on their attribute values. Linear regression provides a straightforward and interpretable method for wine quality estimation.